+++ spec/factories/notifications.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :notification do
+    title { "MyString" }
+    content { "MyText" }
+    user
+    kind { :info }
+    read_at { nil }
+  end
+end